JBERET-220 When a job listener is not found and the batch fails to start, we cannot detect the error using Batch API.

+++ jberet-core/src/main/java/org/jberet/runtime/runner/JobExecutionRunner.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2015 Red Hat, Inc. and/or its affiliates.
+ * Copyright (c) 2012-2016 Red Hat, Inc. and/or its affiliates.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -48,10 +48,11 @@ public final class JobExecutionRunner extends CompositeExecutionRunner<JobContex
             jobExecution.setBatchStatus(BatchStatus.STARTED);
             batchContext.getJobRepository().updateJobExecution(jobExecution, false, false);
         }
-        final JobListener[] jobListeners = createJobListeners();
+        JobListener[] jobListeners = null;
         int i = 0;
 
         try {
+            jobListeners = createJobListeners();
             for (; i < jobListeners.length; i++) {
                 jobListeners[i].beforeJob();
             }
@@ -64,18 +65,22 @@ public final class JobExecutionRunner extends CompositeExecutionRunner<JobContex
         } catch (final Throwable e) {
             BatchLogger.LOGGER.failToRunJob(e, job.getId(), "", job);
             jobExecution.setBatchStatus(BatchStatus.FAILED);
+            if (jobListeners == null) {
+                jobExecution.setExitStatus(e.toString());
+            }
         } finally {
-            for (i = 0; i < jobListeners.length; i++) {
-                try {
-                    jobListeners[i].afterJob();
-                } catch (final Throwable e) {
-                    BatchLogger.LOGGER.failToRunJob(e, job.getId(), "", jobListeners[i]);
-                    jobExecution.setBatchStatus(BatchStatus.FAILED);
+            if (jobListeners != null && jobListeners.length > 0) {
+                for (i = 0; i < jobListeners.length; i++) {
+                    try {
+                        jobListeners[i].afterJob();
+                    } catch (final Throwable e) {
+                        BatchLogger.LOGGER.failToRunJob(e, job.getId(), "", jobListeners[i]);
+                        jobExecution.setBatchStatus(BatchStatus.FAILED);
+                    }
                 }
+                batchContext.destroyArtifact(jobListeners);
             }
         }
-
-        batchContext.destroyArtifact(jobListeners);
 
         boolean saveJobParameters = false;
         switch (jobExecution.getBatchStatus()) {
